Abstract out console log messages for iOS
Apple does not like, and has rejected apps, solely based on console
logging.  For the sake of not having to alter the code every time and
for beginners, I suggest that we make it only log if in debug mode.  I
have not done this for the Android version, but this would probably be
just a useful.